Fix: Replaces ambiguous master ref when appdir is set
When appdir is set, the git push was behaving differently. It attempted
to push to :master, which works when a repository is already set up on
Heroku's side. However, on initial app creation (or first git push), we
need to explicitly push to refs/heads/master or our git push will
complain. We were already doing this for the non-appdir code path.

Fixes #11