Replace deprecated setOutput calls
Should fix #156, `setOutput` was [deprecated][1] and will
be disabled May 31st, 2023 which will lead to this action
breaking or at least no longer outputing whether the deploy
failed or succeeded.

I used the following as reference: [GitHub Docs on Workflow Commands][2].

I would have used `process.env.GITHUB_OUTPUT` directly if it
weren't for this example provided. Might still be fine to do this
but it looks like this is roughly how the `@actions/core` package
[does this internally][3].

They [wrap a lot more logic in an internal function][4] that sadly can't
(or shouldn't be) imported which is a bit of a bummer considering
how simple this operation *should* be. But hey that's Node for ya.

[1]: https://github.blog/changelog/2022-10-11-github-actions-deprecating-save-state-and-set-output-commands/
[2]: https://docs.github.com/en/actions/using-workflows/workflow-commands-for-github-actions#sending-values-to-the-pre-and-post-actions
[3]: https://github.com/actions/toolkit/blob/409d616a6e7610097d47caed40057fc21b43da99/packages/core/src/core.ts#L192-L199
[4]: https://github.com/actions/toolkit/blob/409d616a6e7610097d47caed40057fc21b43da99/packages/core/src/file-command.ts#L11-L25